wip/fix versity integration tests with westend
